+++ .gitignore
@@ -4,3 +4,5 @@ game-engine/coverage/
 .idea/
 
 server/node_modules/
+
+server/dist/
