+++ server/.env
@@ -1,2 +1,2 @@
 PORT=3000
-MODE=PROD
+NODE_ENV=production
