+++ server/src/Player.ts
@@ -1,7 +1,4 @@
-
 interface Player {
-
     id : string
     state : PlayerState
-
 }
