+++ server/utils/generateCode.ts
@@ -1,0 +1,3 @@
+export default function generateCode() :number{
+    return Math.round(Math.random() * 10000)
+}
